Trully research is needed
I did a lot of research and created test files and mini project
I used reserch from Stackoverflow to practice in this mini projects and
files.
I finally got an answer that i need to study in order to understand
And be a ble to replicate on my own.
RESEARCH AND SETTING UP SIDE MINI PROJECT BASED ON THE ACTUAL PROJECT'S
RESEARCH IS AMAZIN AND EYE OPENING.
you learn from experts in the field